+++ go.mod
@@ -1,0 +1,9 @@
+module github.com/XuHuaiyu/tpccgen
+
+go 1.14
+
+require (
+	github.com/go-sql-driver/mysql v1.5.0
+	github.com/spf13/cobra v0.0.6
+	golang.org/x/sync v0.0.0-20181221193216-37e7f081c4d4
+)
